+++ plugin/Terrain/SurfaceSlopeComponent.cs
@@ -30,8 +30,9 @@ namespace groundhog
             pManager.AddMeshParameter("Mesh Faces", "F", "The sub mesh faces (for coloring)", GH_ParamAccess.list);
             pManager.AddPointParameter("Face Centers", "C", "The centers of each mesh face (for vector previews)", GH_ParamAccess.list);
             pManager.AddVectorParameter("Face Vectors", "V", "The direction to the lowest points of each face", GH_ParamAccess.list);
-            pManager.AddNumberParameter("Face Slopes °", "A", "The slope of each mesh face, as an angle", GH_ParamAccess.list);
-            pManager.AddNumberParameter("Face Slopes %", "P", "The slope of each mesh face, as a percentile", GH_ParamAccess.list);
+            pManager.AddNumberParameter("Face Slopes °", "A", "The slope of each mesh face, as the angle of inline", GH_ParamAccess.list);
+            pManager.AddNumberParameter("Face Slopes %", "P", "The slope of each mesh face, as a percentage", GH_ParamAccess.list);
+            pManager.AddNumberParameter("Face Slopes :", "P", "The slope of each mesh face, as the denominator of a ratio (i.e. 1:x)", GH_ParamAccess.list);
         }
 
         protected override void GroundHogSolveInstance(IGH_DataAccess DA)
@@ -49,12 +50,19 @@ namespace groundhog
             var subDirections = TerrainCalculations.GetDirections(subMeshes, subCentres);
             var subAngles = GetAngles(MESH);
 
-            // Calculate ratios from angles
+            // Calculate perentage from angles
             var subPercentiles= new List<double>();
             foreach (var angle in subAngles)
             {
                 var radians = Math.PI * angle / 180.0;
                 subPercentiles.Add(Math.Tan(radians) * 100);
+            }
+
+            // Calculate ratio from percentiles
+            var subRatioDenominators = new List<double>();
+            foreach (var percentage in subPercentiles)
+            {
+                subRatioDenominators.Add(100 / percentage);
             }
 
             // Assign variables to output parameters
@@ -63,6 +71,7 @@ namespace groundhog
             DA.SetDataList(2, subDirections);
             DA.SetDataList(3, subAngles);
             DA.SetDataList(4, subPercentiles);
+            DA.SetDataList(5, subRatioDenominators);
         }
 
         private List<double> GetAngles(Mesh MESH)